// src/theme/extended-tags.generated.ts
// This file is auto-generated by scripts/migrate-theme.ts
export type ExtendedTagName =
  | "warningText"
  | "warningBackground"
  | "warningBorder"
  | "informationText"
  | "informationBackground"
  | "informationBorder"
  | "successText"
  | "successBackground"
  | "successBorder"
  | "defaultText"
  | "defaultBackground"
  | "defaultBorder"
  | "errorText"
  | "errorBackground"
  | "errorBorder"
  | "blueTagText"
  | "blueTagBackground"
  | "blueTagBorder"
  | "cyanTagText"
  | "cyanTagBackground"
  | "cyanTagBorder"
  | "geekblueTagText"
  | "geekblueTagBackground"
  | "geekblueTagBorder"
  | "goldTagText"
  | "goldTagBackground"
  | "goldTagBorder"
  | "greenTagText"
  | "greenTagBackground"
  | "greenTagBorder"
  | "limeTagText"
  | "limeTagBackground"
  | "limeTagBorder"
  | "magentaTagText"
  | "magentaTagBackground"
  | "magentaTagBorder"
  | "orangeTagText"
  | "orangeTagBackground"
  | "orangeTagBorder"
  | "purpleTagText"
  | "purpleTagBackground"
  | "purpleTagBorder"
  | "redTagText"
  | "redTagBackground"
  | "redTagBorder"
  | "volcanoTagText"
  | "volcanoTagBackground"
  | "volcanoTagBorder";